+++ src/java/net/jpountz/lz4/LZ4HCJNICompressor.java
@@ -28,10 +28,7 @@ final class LZ4HCJNICompressor extends LZ4Compressor {
   public int compress(byte[] src, int srcOff, int srcLen, byte[] dest, int destOff, int maxDestLen) {
     checkRange(src, srcOff, srcLen);
     checkRange(dest, destOff, maxDestLen);
-    if (maxDestLen < maxCompressedLength(srcLen)) {
-      throw new IllegalArgumentException("This compressor does not support output buffers whose size is < maxCompressedLength(srcLen)");
-    }
-    final int result = LZ4JNI.LZ4_compressHC(src, srcOff, srcLen, dest, destOff);
+    final int result = LZ4JNI.LZ4_compressHC(src, srcOff, srcLen, dest, destOff, maxDestLen);
     if (result <= 0) {
       throw new LZ4Exception();
     }